--- User/route.py
@@ -1,4 +1,0 @@
-
-@app.route('/login')
-def login_method():
-   return
