+++ app/__init__.py
@@ -1,0 +1,9 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+
+app = Flask(__name__)
+app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql://flask:bhuvnesh@curesee.in/flask'
+app.config['SECRET_KEY'] = 'Dealord'
+db = SQLAlchemy(app)
+
+from app.User import *
